+++ internal/dev_server/model/store.go
@@ -46,23 +46,4 @@ func StoreMiddleware(store Store) mux.MiddlewareFunc {
 }
 
 var ErrNotFound = errors.New("not found")
-
-type Error struct {
-	err     error
-	message string
-}
-
-func (e Error) Error() string {
-	return e.message
-}
-
-func (e Error) Unwrap() error {
-	return e.err
-}
-
-func NewError(message string) error {
-	return errors.WithStack(Error{
-		err:     errors.New(message),
-		message: message,
-	})
-}
+var ErrAlreadyExists = errors.New("already exists")
